Update slot property name to avoid incorrect type
Simply renamed the "data" variable in the "each" loop to "dataItem" to avoid producing incorrect typescript definition on the default slot's property "data". As "data" is declared also at the top as the main component's property, with a type of Array<any>, the slot's "data" property was also producing the same type as Array<any>, where it should just be "any" as it's a single data item.
As a work-around, I renamed the local loop variable to "dataItem" so Sveld treats it as a different variable from the "data" at the top, and produces the correct type definition.

It is a bug from Sveld, but for now this workaround works.